Refactor commands structure.
Split commands by "category" and remove all of them from main
entrypoint to make it clean and maintainable (hope so).
