+++ commands/config.py
@@ -1,0 +1,5 @@
+CHARADA_ENDPOINT = "https://us-central1-kivson.cloudfunctions.net/charada-aleatoria"
+COACH_ENDPOINT = "http://api.forismatic.com/api/1.0/?method=getQuote&key=457653&format=json&lang=en"
+YT_RESULTS_ENDPOINT = "http://www.youtube.com/results?"
+YT_WATCH_ENDPOINT = "http://www.youtube.com/watch?v="
+HOROSCOPO_ENDPOINT = "http://babi.hefesto.io/signo/{}/dia"
