+++ commands/utils.py
@@ -1,0 +1,3 @@
+from discord.ext import commands
+
+bot = commands.Bot(command_prefix=".")
